+++ src/Raytracing.jl
@@ -7,7 +7,7 @@
 module Raytracing
 
 using Base: Bool
-using Colors, Images, ImageIO, FileIO
+using Colors, Images, ImageIO, FileIO, Polynomials
 using ColorTypes:RGB
 using LinearAlgebra, StaticArrays
 using Printf, ProgressBars
@@ -20,7 +20,7 @@ import LinearAlgebra.:⋅; import LinearAlgebra.:×
 export BLACK, WHITE, to_RGB, HDRimage, Parameters
 export Point, Vec, Normal,VEC_X, VEC_Y, VEC_Z, Transformation
 export Ray, OrthogonalCamera, PerspectiveCamera, ImageTracer
-export Shape, PointLight, Sphere, Plane, Vec2d, HitRecord, World
+export Shape, PointLight, Sphere, Plane, Torus, Vec2d, HitRecord, World
 export Pigment, UniformPigment, CheckeredPigment, ImagePigment
 export BRDF, DiffuseBRDF, SpecularBRDF, Material
 export Renderer, OnOffRenderer, FlatRenderer, PathTracer, PointLightRenderer
@@ -43,7 +43,7 @@ export sphere_point_to_uv, sphere_normal
 export plane_point_to_uv, plane_normal
 export is_point_visible, quick_ray_intersection
 # from Demo.jl
-export first_world, second_world
+export first_world, second_world, select_world
 export demo, demo_animation
 # from Renderers.jl
 export choose_renderer, call