+++ test/test_Renderers.jl
@@ -5,9 +5,6 @@
 # Copyright © 2021 Matteo Foglieni and Riccardo Gervasoni
 #
 
-
-
-
 @testset "test_OnOffRenderer" begin
      sphere = Sphere(translation(Vec(2, 0, 0)) * scaling(Vec(0.2, 0.2, 0.2)),
                     Material(DiffuseBRDF(UniformPigment(WHITE))))
@@ -97,6 +94,16 @@ end
 end
 
 @testset "test_PointLightTracer" begin
+<<<<<<< HEAD
+     @testset "test_quick_ray_intersection" begin
+          world = World()
+
+          sphere1 = Sphere(translation(VEC_X * 2.0))
+          sphere2 = Sphere(translation(VEC_X * 8.0))
+          add_shape!(world, sphere1)
+          add_shape!(world, sphere2)
+
+=======
      @testset "quick_ray_intersection_sphere" begin
           world = World()
 
@@ -129,6 +136,7 @@ end
           add_shape!(world, plane2)
 
           @test is_point_visible(world, Point(-10.0, 0.0, 0.0), Point(0.0, 0.0, 0.0))
+>>>>>>> ddb98273ccf586f2bf52ed0b928d7a4851a59b77
           @test !is_point_visible(world, Point(10.0, 0.0, 0.0), Point(0.0, 0.0, 0.0))
           @test !is_point_visible(world, Point(5.0, 0.0, 0.0), Point(0.0, 0.0, 0.0))
           @test is_point_visible(world, Point(5.0, 0.0, 0.0), Point(4.0, 0.0, 0.0))