+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # HEAD
+<<<<<<< HEAD
+- Added antialiasing algorithm
 
 
 - Added point-light tracing algorithm, see PR[#17](https://github.com/cosmofico97/Raytracing/pull/17)
@@ -6,12 +8,24 @@
 # VERSION 0.3.0
 
 - Added antialiasing algorithm, see PR[#16](https://github.com/cosmofico97/Raytracing/pull/16)
+=======
+- Added antialiasing algorithm, as default uses 9x9 rays in a pixel, missing command line option
+>>>>>>> antialiasing
+
+<<<<<<< HEAD
+<<<<<<< HEAD
+=======
+=======
+>>>>>>> ddb98273ccf586f2bf52ed0b928d7a4851a59b77
+
+# VERSION 0.2.0
+>>>>>>> pointlighttracing
 
 
 # VERSION 0.2.0
-
 - Modified `demo` function: now two possible world versions can be rendered
 - New struct for generate a better image considering lights and angles (`PathTracer`) and tests
+- New shape `Torus` with relative struct and functions
 - Implemented scalar and vector products between two `Normal` and between `Normal` and `Vec`
 - New struct `SpecularBRDF`
 - Implemented PCG new struct (`PCG`) and algorithm for random numbers
@@ -21,8 +35,9 @@
 - New struct for implement the material of a Shape (`Material`) and tests
 - New structs for BRDF implementtion (`BRDF`, `DiffuseBRDF`) and tests
 - New structs for pigment implementation (`Pigment`, `UniformPigment`, `CheckeredPigment`, `ImagePigment`) and tests
-- Modified operation between `Transformation` and `Normal`, `Vec` and `Point` for time optimization
-- Added test for World, modified test for Sphere
+- New struct from Shape: `Plane`
+- Modified operations and constructors between `Transformation` and `Normal`, `Vec` and `Point` for time optimization
+- Added test for `World`, modified test for `Sphere`
 - Added feature: now can create an animation
 - New interface from command line
 - Added test for camera orientation ([PR#8](https://github.com/cosmofico97/Raytracing/pull/8#issue-631504956))
@@ -31,6 +46,5 @@
 - Implemented structs: `Shape`, `Sphere`, `World` for a first image creation
 
 # VERSION 0.1.0
-
 - Added feature: now can convert from a .pfm image to .png and .tiff one 
 - Completed function for reading and writing a .pfm image file format